Finished for the user to add the extra details for him!

+++ src/main/java/com/kara/studentscareer/bachelorpapel/service/ExperienceServiceImpl.java
@@ -1,49 +1,34 @@
 package com.kara.studentscareer.bachelorpapel.service;
 
 import com.kara.studentscareer.bachelorpapel.entity.Experience;
+import com.kara.studentscareer.bachelorpapel.entity.User;
 import com.kara.studentscareer.bachelorpapel.repository.ExperienceRepository;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import javax.persistence.EntityNotFoundException;
-import java.util.List;
 import java.util.Optional;
+
 @Service
-public class ExperienceServiceImpl implements ExperienceService {
+public class ExperienceServiceImpl implements ExperienceService{
 
+
+    @Autowired
     private ExperienceRepository experienceRepository;
-
     @Override
-    public Experience createExperience(Experience experience) {
-        return experienceRepository.save(experience);
+    public void addNewExperience(String company, String position, String startDate, String endDate, User loggedInUser) {
+        Experience newExperience=new Experience(company,position,startDate,endDate,loggedInUser);
+        experienceRepository.save(newExperience);
     }
 
     @Override
-    public List<Experience> findAllExperiences() {
-        return experienceRepository.findAll();
-    }
-
-    @Override
-    public Optional<Experience> findExperienceById(Integer id) {
-        return experienceRepository.findById(id);
-    }
-
-    @Override
-    public Experience updateExperience(Integer id, Experience updatedExperience) {
-        Optional<Experience> isThereExperience=findExperienceById(id);
-        if(isThereExperience.isEmpty()){
-            throw new EntityNotFoundException("There is no experience with this ID: "+id);
+    public void deleteExperience(Integer id) {
+        Optional<Experience> experienceToDelete=experienceRepository.findById(id);
+        if(experienceToDelete.isPresent()){
+            experienceRepository.deleteById(id);
         }
-        Experience existingExperience=isThereExperience.get();
-        existingExperience.setCompany(updatedExperience.getCompany());
-        existingExperience.setPosition(updatedExperience.getPosition());
-        existingExperience.setStartDate(updatedExperience.getStartDate());
-        existingExperience.setEndDate(updatedExperience.getEndDate());
-        experienceRepository.save(existingExperience);
-        return existingExperience;
-    }
-
-    @Override
-    public void deleteExperienceById(Integer id) {
-        experienceRepository.deleteById(id);
+        else {
+            throw new EntityNotFoundException("Your emails doesnt exist!");
+        }
     }
 }
